refactoring extractRPC helper. changign underscore to lodash.
